fix(studio): hide opaque H3 generation models
